These changes account for the change in the rules that specifies that the preloads have to come from the 4 samples (2 alliance, 2 yellow) of the human player's starting supply.
This changes strategy pretty significantly.

It seems like in a vast majority of cases, you want the Net bot to start with a sample preload and the Observation bot to start with a specimen. If you don't do it this way, then the Observation bot has to score the preload specimen and bring an alliance sample to the HP before they can start making a specimen.

Maybe there is some world where you want the Net bot to start with a specimen and the Observation bot to start with a Sample. This would only really work with coordinated bots though. It might let you quick score in the basket.